requirements.txt
albumentations==1.4.18
blobfile==2.0.2
clip==0.2.0
einops==0.8.0
kornia==0.6.12
ldm==0.1.3
loralib==0.1.2
lpips==0.1.4
natsort==8.4.0
numpy==1.21.6
omegaconf==2.3.0
open_clip_torch==2.24.0
opencv_python==4.8.0.74
opencv_python_headless==4.5.5.64
Pillow==9.5.0
Pillow==9.4.0
Pillow==11.0.0
pytorch_lightning==1.9.5
scipy==1.7.3
taming_transformers==0.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
